fix: message not recorded when conversation is not focussed
